+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -9,25 +9,25 @@ public class StatsServiceTest {
     @Test
     void shouldCalculateSum() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 180;
-        int actual = service.sum(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 180;
+        long actual = service.sum(sales);
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
     void shouldCalculateAverage() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 15;
-        int actual = service.average(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 15;
+        long actual = service.average(sales);
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
     void shouldFindMaxSalesMonth() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 8;
         int actual = service.maxSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -36,7 +36,7 @@ public class StatsServiceTest {
     @Test
     void shouldFindMinSalesMonth() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 9;
         int actual = service.minSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -45,7 +45,7 @@ public class StatsServiceTest {
     @Test
     void shouldCalculateBelowAverageMonths() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
         int actual = service.belowAverage(sales);
         Assertions.assertEquals(expected, actual);
@@ -54,7 +54,7 @@ public class StatsServiceTest {
     @Test
     void shouldCalculateAboveAverageMonths() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
         int actual = service.aboveAverage(sales);
         Assertions.assertEquals(expected, actual);
